% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/empatica.R
\docType{package}
\name{empatica}
\alias{empatica}
\alias{empatica-package}
\title{Read Physiologic Signal Data Recorded with a Empatica Device.}
\description{
The empatica package contains functions for reading data from an
Empatica device.
}

