% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{organise_data}
\alias{organise_data}
\title{Create a data structure containing data and sampling rate from the Empatica data.}
\usage{
organise_data(data, samplingrate = NULL)
}
\arguments{
\item{data}{A data frame with the data.}

\item{samplingrate}{The sampling rate in Hz of the signal, if applicable. Default is \code{NULL}.}
}
\value{
The header information (timestamp).
}
\description{
Create a data structure containing data and sampling rate from the Empatica data.
}
\keyword{internal}

