% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_empatica.R
\name{read.empatica}
\alias{read.empatica}
\title{Read data from an Empatica device.}
\usage{
read.empatica(filename, header.only = FALSE)
}
\arguments{
\item{filename}{The name of the zip file containing the Empatica data}

\item{header.only}{Should only the header information, but no signal data, be read. Boolean. Default is FALSE.}
}
\value{
A recording structure (a list) containing the data.
}
\description{
Read data from an Empatica device.
}

