% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_empatica_events}
\alias{read_empatica_events}
\title{Read Empatica events (button presses)}
\usage{
read_empatica_events(f)
}
\arguments{
\item{f}{The name of the file containing the events (\code{tags.csv}).}
}
\value{
A data frame with the events
}
\description{
Read Empatica events (button presses)
}
\keyword{internal}

