% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_empatica_ibi}
\alias{read_empatica_ibi}
\title{Read interbeat interval (IBI) data in the empatica format.}
\usage{
read_empatica_ibi(filename, signal_names = c("time", "ibi"))
}
\arguments{
\item{filename}{The name of the Empatica signal file with IBI data (\code{ibi.csv}).}

\item{signal_names}{A list containing names for the data. Default is \code{signal_names = c("time", "ibi")}.}
}
\value{
The header information (timestamp).
}
\description{
Read interbeat interval (IBI) data in the empatica format.
}
\keyword{internal}

