% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_header}
\alias{read_header}
\title{Read header information from an Empatica signal file.}
\usage{
read_header(filename)
}
\arguments{
\item{filename}{The name of an Empatica signal file.}
}
\value{
The header information (timestamp).
}
\description{
Read header information from an Empatica signal file.
}
\keyword{internal}

